% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCooc.R
\name{CalGeneRankings}
\alias{CalGeneRankings}
\title{Calculate Gene Rankings}
\usage{
CalGeneRankings(
  object,
  min.expr.cells = 20,
  min.expr.pct = 0.001,
  nfeatures = 8000,
  features = NULL,
  ndim.mca = 30,
  use.variable.features = TRUE,
  batch.name = NULL,
  module.source = "GeneCooc",
  assay = "RNA"
)
}
\arguments{
\item{object}{A Seurat object.}

\item{min.expr.cells}{The minimum number of cells in which a gene must be expressed
to be considered. Defaults to 20. If min.expr.pct is specified, min.expr.cells is
adjusted to be the minimal between min.expr.cells and ncol(object)*min.expr.pct.}

\item{min.expr.pct}{The minimum percentage of cells in which a gene must be expressed
to be considered. Defaults to 0.001.}

\item{nfeatures}{The number of features to choose when finding variable features.
Defaults to 8000.}

\item{features}{An optional vector of specific features to use for the analysis.
If NULL, features will be automatically selected based on expression criteria.
Defaults to NULL.}

\item{ndim.mca}{The number of dimensions to consider in the MCA. Defaults to 30.}

\item{use.variable.features}{Logical, indicating whether to only consider variable
features. If TRUE, the function searches for variable features using the \code{FindVariableFeatures}
function from Seurat. Defaults to TRUE.}

\item{batch.name}{An optional character string representing the batch variable
to account for batch effects when selecting variable features. If NULL, batch effects are not considered during
the variable features selection. Defaults to NULL.}

\item{module.source}{A character string indicating where to save results of \code{GeneCooc}. Default is "GeneCooc".}

\item{assay}{A character string specifying which assay to use from the Seurat
object. Defaults is "RNA".}
}
\value{
A modified Seurat object inclusive of the \code{features.use} and \code{gene.rankings} in its
\code{misc} slot in the list named by \code{module.source}.
}
\description{
This function calculates gene rankings using Multiple Correspondence Analysis (MCA) implemented in CelliD.
}
