% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCooc.R
\name{CalModuleScore}
\alias{CalModuleScore}
\title{Calculate Module Scores for Cells in a Seurat Object}
\usage{
CalModuleScore(object, ndim.mca = 30, module.source = "GeneCooc")
}
\arguments{
\item{object}{A Seurat object containing gene module data in its \code{misc} slot,
and embeddings from MCA.}

\item{ndim.mca}{The number of dimensions to use from the multiple correspondence analysis
(MCA) embeddings for calculating the module scores. Default is 30 dimensions.}

\item{module.source}{A character string indicating where to load tmp results and save final results
of \code{GeneCooc}. Default is "GeneCooc".}
}
\value{
Seurat object updated with a new assay slot named \code{module.score} containing the
calculated scores for each module and cell.

TODO calculate module scores for a given gene module list
}
\description{
Computes scores for gene modules in each cell based on multiple correspondence analysis (MCA)
embedding and the association of genes to modules. It first filters out genes that have been
trimmed, then calculates cell-to-gene distances, followed by aggregating these distances to
compute module scores for each cell.
}
