% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{FetchAffinityMatrix}
\alias{FetchAffinityMatrix}
\title{Fetch Affinity Matrix From Seurat Object}
\usage{
FetchAffinityMatrix(object, module.source = "GeneCooc")
}
\arguments{
\item{object}{A Seurat object.}

\item{module.source}{A character string indicating where to fetch data. Default is "GeneCooc".}
}
\value{
An affinity matrix associated with the specified key in the Seurat object's
miscellaneous data slot or another designated slot where such data is stored.
}
\description{
Retrieves an affinity matrix from a Seurat object based on the specified module source.
}
