% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{FetchModuleDR}
\alias{FetchModuleDR}
\title{Fetch Module Dimensionality Reduction Data}
\usage{
FetchModuleDR(
  object,
  reduction = "FDG",
  slot = "all",
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object containing module dimensionality reduction information.}

\item{reduction}{A character string indicating which type of dimensionality reduction data
to fetch. Default is "FDG" for Force-Directed Graph data.}

\item{slot}{A character string specifying the module slot to retrieve data from.
Default is "all" which would typically denote the combined results for all modules.}

\item{module.source}{A character string specifying the key under which the module
dimensionality reduction data is stored within the Seurat object's \code{misc} slot. Default is "GeneCooc".}
}
\value{
Returns a data structure containing the dimensionality reduction data.
The exact structure will depend on how the data is stored within the provided Seurat object.
}
\description{
This function retrieves dimensionality reduction data for gene modules
from a specified slot within a given Seurat object.
}
