% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{FetchModuleList}
\alias{FetchModuleList}
\title{Fetch Gene Module List From Seurat Object}
\usage{
FetchModuleList(
  object,
  module.source = "GeneCooc",
  module.type = "both",
  module.size = 10
)
}
\arguments{
\item{object}{A Seurat object.}

\item{module.source}{A string specifying the key or named list within the Seurat object from
which to retrieve the module information. The default value is "GeneCooc".}

\item{module.type}{A string specifying the type of gene modules to include in the output list:
"both" for both major and minor modules, "major" for major modules only, or "minor" for minor modules
only. The default value is "both".}

\item{module.size}{Only returns the modules with sizes more than \code{module.size}. Default: 10.}
}
\value{
A list of gene names associated with the specified module source and type in the Seurat
object. The list is organized based on the major and minor module categories, if applicable.
}
\description{
Retrieves a gene module list from a Seurat object based on the specified module source and type.
This function returns a list of gene names associated with different modules within the provided
Seurat object.
}
