% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCooc.R
\name{FindModules}
\alias{FindModules}
\title{Find Gene Modules}
\usage{
FindModules(
  object,
  k = 50,
  resolution = 0.1,
  min.module.size = 10,
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object containing gene co-occurrence data in an affinity matrix within
its \code{misc} slot.}

\item{k}{Integer. Specifies the number of nearest neighbors to be used in the construction of
the SNN (shared nearest neighbor) graph.}

\item{resolution}{A double. Resolution parameter for the Louvain clustering algorithm. Controls
the granularity of the clustering.}

\item{min.module.size}{Integer. The minimum size of the modules to be detected in the dynamic
tree cut.}

\item{module.source}{A character string indicating where to load tmp results and save final results
of \code{GeneCooc}. Default is "GeneCooc".}
}
\value{
Seurat object with the gene module information computed. The results will be written
into the object's \code{misc} slot under \code{GeneCooc}.
}
\description{
This function identifies gene modules based on an affinity matrix within a given Seurat object.
It uses the Louvain algorithm for primary clustering and a dynamic tree cut for sub-module detection.
}
