% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcl_hsa}
\alias{hcl_hsa}
\title{A data set containing marker genes identified in Human Cell Lanscape (HCL)
in human symbols}
\format{
A data frame with 20400 rows and 2 variables:
\describe{
\item{ont}{description of gene set}
\item{gene}{human gene symbols}
...
}
}
\source{
\url{https://doi.org/10.1038/s41586-020-2157-4}
}
\usage{
hcl_hsa
}
\description{
A data set containing marker genes identified in Human Cell Lanscape (HCL)
in human symbols
}
\keyword{datasets}
