% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mca_hsa}
\alias{mca_hsa}
\title{A data set containing marker genes identified in Mouse Cell Atlas (MCA)
in human symbols}
\format{
A data frame with 32870 rows and 2 variables:
\describe{
\item{ont}{description of gene set}
\item{gene}{human gene symbols}
...
}
}
\source{
\url{https://doi.org/10.1016/j.cell.2018.02.001}
}
\usage{
mca_hsa
}
\description{
A data set containing marker genes identified in Mouse Cell Atlas (MCA)
in human symbols
}
\keyword{datasets}
