% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualization.R
\name{ModuleDimPlot}
\alias{ModuleDimPlot}
\title{Visualize modular gene interactions using a dimensionality reduction plot}
\usage{
ModuleDimPlot(
  object,
  reduction = "FDG",
  group.by = "major.modules",
  slot = "all",
  module.source = "GeneCooc",
  pt.size = 1,
  pt.border = NULL,
  pt.colors = NULL,
  line.color = "grey",
  line.alpha = 0.01
)
}
\arguments{
\item{object}{A Seurat object containing module dimensionality reduction data.}

\item{reduction}{A character string specifying which dimensionality reduction technique
was used. Default is "FDG" for Force-Directed Graph.}

\item{group.by}{A character string indicating the field in the \code{data.nodes}
dataframe to use for coloring the points. Default is "major.modules".}

\item{slot}{A character string indicating the specific module slot to use for the plot.
Default is "all".}

\item{module.source}{A character string specifying the source of the module data within
the Seurat object. Defaults to "GeneCooc".}

\item{pt.size}{A numeric value specifying the size of the points in the plot. Default is 1.}

\item{pt.border}{A numeric value specifying the size of the point borders. If NULL,
no borders will be added to the points. Default is NULL.}

\item{pt.colors}{A character vector of colors to use for the points. If NULL,
a predefined color scale is used. Default is NULL.}

\item{line.color}{A character string specifying the color of the lines connecting points.
Default is "grey".}

\item{line.alpha}{A numeric value specifying the transparency of the lines connecting
points. Default is 0.01.}
}
\value{
A \code{ggplot} object representing the dimensionality reduction plot.
}
\description{
This function creates a plot displaying the connections and relationships between genes
in major modules, visualized according to a chosen dimensionality reduction technique.
}
