% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualization.R
\name{ModuleHeatmap}
\alias{ModuleHeatmap}
\title{Create a Heatmap to Visualize Gene Module Affinity}
\usage{
ModuleHeatmap(
  object,
  module.use = "all",
  module.type = "major",
  module.size = 10,
  colors = NULL,
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object.}

\item{module.use}{A character vector specifying which modules to use for the heatmap.
If "all", all available modules are selected. Default is "all".}

\item{module.type}{A character string specifying whether to use "major" or "minor"
modules for heatmap visualization. Default is "major".}

\item{module.size}{An integer indicating the minimum number of genes a module must contain
to be included in the analysis. Default is 10.}

\item{colors}{A character vector of colors to use in the heatmap column annotation. If NULL,
a default palette (ggsci::pal_d3) is provided. Default is NULL.}

\item{module.source}{A character string specifying the source of the module data within
the Seurat object. Defaults to "GeneCooc".}
}
\value{
Heatmap plot from the ComplexHeatmap package.
}
\description{
This function generates a heatmap to visualize gene interactions within specified modules
based on an affinity matrix. The resulting heatmap can represent either major or minor
gene modules and allows customization of color schemes.
}
