% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ModuleUMAPPlot}
\alias{ModuleUMAPPlot}
\title{Visualize Gene Modules on UMAP Plot}
\usage{
ModuleUMAPPlot(
  object,
  pt.size = 1,
  pt.alpha = 1,
  text.size = 3,
  genes.highlight = NULL,
  colors = NULL,
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object that contains UMAP coordinates and gene module data.}

\item{pt.size}{Numeric size of the points in the UMAP plot.}

\item{pt.alpha}{Numeric value representing the transparency of the points in the UMAP plot.
Values range from 0 (transparent) to 1 (opaque).}

\item{text.size}{Numeric size of the text labels for highlighted genes.}

\item{genes.highlight}{A vector of gene names to highlight on the UMAP plot. Highlighted
genes will be marked with different point shapes and labeled with text.}

\item{colors}{A named vector of colors to be used for coloring the gene modules. The names
should correspond to the unique module names.}

\item{module.source}{The name of the source from which to fetch the gene module data.
Defaults to "GeneCooc".}
}
\value{
A ggplot2 object containing the UMAP plot.
}
\description{
Creates a UMAP plot to display gene modules as colored points. The plot can be customized with
different point sizes, transparencies, and an option to highlight specific genes.
}
