% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{RescueModules}
\alias{RescueModules}
\title{Rescue Specified Modules}
\usage{
RescueModules(object, modules, module.source = "GeneCooc")
}
\arguments{
\item{object}{A Seurat object containing various types of data, including gene module information.}

\item{modules}{A character vector of minor module names (\code{minor.module.full}) to be marked as kept
within the Seurat
object.}

\item{module.source}{A string specifying the name of the list key within the Seurat object's
\verb{@misc} slot from which gene module information should be fetched and updated.
The default value is "GeneCooc".}
}
\value{
The modified Seurat object with the specified subset of modules marked as "kept".
}
\description{
This function "rescues" certain gene modules in a Seurat object by marking them as kept,
effectively including them for further analysis or consideration. The specified modules
are altered to have their \code{is.kept} status set to \code{TRUE} within the module source data.
}
