% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualization.R
\name{RunModuleFDG}
\alias{RunModuleFDG}
\title{Run Force-Directed Graph (FDG) on Affinity Matrix}
\usage{
RunModuleFDG(
  object,
  module.source = "GeneCooc",
  module.size = 10,
  weight.cutoff = 0.1,
  on.module = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{module.source}{A character string indicating the source of the module data within the
Seurat object. Default is "GeneCooc".}

\item{module.size}{An integer indicating the minimum number of genes a module must contain
to be included in the analysis. Default is 10.}

\item{weight.cutoff}{A numeric threshold for the values in the affinity matrix. Values with
below this cutoff will be discarded. Default is 0.1.}

\item{on.module}{An optional character string specifying a particular module to analyze.
If NULL, all modules are considered. Default is NULL.}
}
\value{
A modified Seurat object that includes the force-directed graph layout coordinates
and adjacency data in its \code{misc} slot under a list named by \code{module.source}.
}
\description{
This function constructs a force-directed graph to visualize the relationships between genes
in selected modules. It uses the igraph package to create an undirected weighted graph and
determines the layout using the Fruchterman-Reingold algorithm.
}
