% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCooc.R
\name{RunModuleUMAP}
\alias{RunModuleUMAP}
\title{Run UMAP on Gene Modules}
\usage{
RunModuleUMAP(
  object,
  exclude.trimmed = TRUE,
  supervised = FALSE,
  module.source = "GeneCooc",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing gene module data and an affinity matrix within its
\code{misc} slot.}

\item{exclude.trimmed}{Logical, indicating whether to exclude trimmed genes from UMAP analysis.
If \code{TRUE}, genes that are not marked as kept will be omitted.}

\item{supervised}{Logical, controls the mode of UMAP execution. If \code{FALSE}, the UMAP runs in
unsupervised mode. If \code{TRUE}, it utilizes gene module labels to steer the embedding in a
supervised fashion.}

\item{module.source}{A character string indicating where to load tmp results and save final results
of \code{GeneCooc}. Default is "GeneCooc".}

\item{...}{Additional arguments passed on to the uwot::umap function.}
}
\value{
Seurat object with updated gene module information including UMAP coordinates.
}
\description{
This function runs UMAP (Uniform Manifold Approximation and Projection) on gene modules within
a given Seurat object. It supports both supervised and unsupervised modes.
}
