% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{softmax}
\alias{softmax}
\title{Softmax Function}
\usage{
softmax(x)
}
\arguments{
\item{x}{A numeric vector or array.}
}
\value{
A numeric vector representing the softmax transformation of the input.
}
\description{
The softmax function, also known as the normalized exponential function, is a generalization
of the logistic function that squashes a K-dimensional vector of arbitrary real values
to a K-dimensional vector of real values in the range (0, 1) that add up to 1.
It is commonly used in classification problems.
}
