% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SubModuleUMAPPlot}
\alias{SubModuleUMAPPlot}
\title{Visualize Specific Gene Sub-Modules on UMAP Plot}
\usage{
SubModuleUMAPPlot(
  object,
  pt.size = 1,
  pt.alpha = 1,
  text.size = 3,
  module.use = NULL,
  genes.highlight = NULL,
  colors = NULL,
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object that contains UMAP coordinates and gene module data.}

\item{pt.size}{Numeric size of the points in the UMAP plot.}

\item{pt.alpha}{Numeric value representing the transparency of the points in the UMAP plot.
Values range from 0 (transparent) to 1 (opaque).}

\item{text.size}{Numeric size of the labels for highlighted genes on the UMAP plot.}

\item{module.use}{A character vector indicating which sub-modules should be highlighted on
the UMAP plot. This parameter is required.}

\item{genes.highlight}{A vector of gene names to highlight on the plot.}

\item{colors}{A named vector of colors to be used for the sub-modules. The names should
correspond to the unique names of the sub-modules you wish to color.}

\item{module.source}{The name of the source from which to fetch the gene module data.
Defaults to "GeneCooc".}
}
\value{
A \code{ggplot2} object containing the UMAP plot with distinct visual emphasis on
the selected gene sub-modules.
}
\description{
Generates a UMAP plot that highlights specific gene sub-modules within a Seurat object.
Additional genes may be highlighted with distinct markers and labels for further emphasis.
}
