% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCooc.R
\name{TrimModules}
\alias{TrimModules}
\title{Trim Gene Modules Based on Archetype Analysis}
\usage{
TrimModules(
  object,
  archetype.score.cutoff = 0.5,
  delta.cutoff = 0.2,
  module.source = "GeneCooc"
)
}
\arguments{
\item{object}{A Seurat object containing gene module data and an affinity matrix within its
\code{misc} slot.}

\item{archetype.score.cutoff}{A numeric threshold for the archetype score. Genes with an
archetype score below this cutoff will not be considered archetypes and may be trimmed
accordingly.}

\item{delta.cutoff}{A numeric threshold for the difference in archetype scores. If the difference
between the scores of the most expressive genes (archetypes) is less than this value, they may be
excluded from the module.}

\item{module.source}{A character string indicating where to load tmp results and save final results
of \code{GeneCooc}. Default is "GeneCooc".}
}
\value{
Seurat object with updated sub-module information, including flags to indicate which
genes are kept and which ones are identified as archetypes for their respective sub-modules.
}
\description{
Performs trimming of gene sub-modules within the given object using archetype analysis. The
function evaluates each sub-module based on archetypes and trims them according to specified
cutoff values.
}
