% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_decomposition.R
\name{VarDecompose}
\alias{VarDecompose}
\title{Perform Variance Decomposition}
\usage{
VarDecompose(object, vd.vars, assay = "GeneCooc", cores = 1)
}
\arguments{
\item{object}{A Seurat object containing gene expression data and meta data.}

\item{vd.vars}{A character vector specifying the variables of interest to perform variance
decomposition.}

\item{assay}{A character string indicating the assay in the Seurat object to use for the
variance decomposition. Default is "GeneCooc".}

\item{cores}{An integer value specifying the number of CPU cores to use for parallel
computation. Default is 1.}
}
\value{
A data frame containing the variance decomposition results.
}
\description{
This function performs variance decomposition for a set of variables of interest (\code{vd.vars})
across genes within the provided Seurat object. Variance decomposition determines the
contribution of each variable of interest to the total variance in gene expression. The
function uses a linear mixed-effects model to estimate the variance components. The result is
a table containing the variance contribution for each variable, as well as the residual
variance.
}
